added settings for test

+++ back/src/main/java/ru/dreamteam/entities/PayoutRequestToPlatformEntity.java
@@ -10,8 +10,8 @@ public class PayoutRequestToPlatformEntity implements Json{
     private String currency;
     private String partnerPayoutId;
     private String destinationId;
-    private Integer accountId;
-    private Integer accountGroupId;
+    private Long accountId;
+    private Long accountGroupId;
     private String method;
     private String destination;
     private Extra extra;
